requirements.txt
flask
python-dotenv
line-bot-sdk==3.7.0
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
gunicorn
